int the n_samples

+++ utils/data_loader.py
@@ -98,7 +98,7 @@ class DataLoader:
 
             # maj_len = max(len(c0), len(c1), len(c2))
             maj_len = len(c2)
-            c0 = resample(c0, n_samples=maj_len * 1.5, random_state=FLAGS.random_state)
+            c0 = resample(c0, n_samples=int(maj_len * 1.5), random_state=FLAGS.random_state)
             c1 = resample(c1, n_samples=maj_len, random_state=FLAGS.random_state)
             c2 = resample(c2, n_samples=maj_len, random_state=FLAGS.random_state)
